Extracted FileParser from GridFromFileGenerator. Refactored existing tests and wrote new ones as appropriate.
