--- conways2/test/com/ness/gridSizeLargerThanMaximum.txt
@@ -1,1 +1,0 @@
-57 12
