--- conways2/test/com/ness/gridSizeLessThanZero.txt
@@ -1,1 +1,0 @@
--3 34
