+++ conways2/test/com/ness/incorrectlyFormattedSecondLine.txt
@@ -1,0 +1,2 @@
+1 1
+1,1
