--- conways2/test/com/ness/setup1.txt
@@ -1,1 +1,0 @@
-1 1
